[FIX] approvals_purchase: Prevent creating RFQ if already created
**Problem:**
It's possible to click the "Create RFQ's" button more than once, as the user
may have multiple tabs open or multiple users are viewing the same record. When
this happens, the approval will create or add to an RFQ even if it already did,
and this causes double the intended product quantities.

**Solution:**
The "Create RFQ's" button becomes hidden when purchase_order_count > 0 (i.e.
there are linked POs) so we can perform this check within the button's method
`action_create_purchase_orders` to prevent RFQ generation (or modification).

closes odoo/enterprise#100108

X-original-commit: f76a2cc2d43bf46bf426052e1e1f5a8ae676ac7c
Signed-off-by: Dirk Douglas (dido) <dido@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>