[FIX] appointment: fix add closing days button visibility
Purpose
=======
Fix the "Add Closing Day(s)" button visibility.

Specification
=============
The button should be visible on every resource views (gantt, calendar, list),
and should be invisible on every staff booking views (gantt, calendar, list).

Currently, the button is visible on the list view for staff bookings, it should not.
The button is also not present on the calendar view for resource bookings, it should be.

Task-5079021

closes odoo/enterprise#96027

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>