[FIX] stock_barcode: split correctly quantity sublines same lot
**Steps to reproduce:**
- create a product tracked by lot
- create a lot with a quantity of 2
- create a new sale order
- add two sale order lines, both for a
quantity of 1 of the product
- confirm
- open the picking in barcode
- scan the stock location
- scan the lot
- scan the lot another time
- validate

**Current behavior:**
a backorder is created

**Expected behavior:**
No back order should be created

**Cause of the issue:**
After scanning the lot for the first time we have
the following situation:
two lines :
- one with a quantity of 1, qty_done of 1 and
reserved_uom_qty of 1.
- one with a quantity of 1, qty_done of 0 and
reserved_uom_qty of 1.
both lined grouped in a parent line with quantity
of 1 qty_done of 1 and reserved_uom_qty of 2.
All of this is correct.

when scanning the lot for the second time:
_findLine iterates through the lines to select
the right line to use.
 _findLine calls _lineIsNotComplete on the first line
to check if it's complete (this first line is complete).
https://github.com/odoo/enterprise/blob/58d55868750b827a9d5ebd8b4ab2cc23c4445eca/stock_barcode/static/src/models/barcode_model.js#L1684
But _lineIsNotComplete will actually do the check on the
parent line (which is not complete), so the return value
will be true.
https://github.com/odoo/enterprise/blob/58d55868750b827a9d5ebd8b4ab2cc23c4445eca/stock_barcode/static/src/models/barcode_picking_model.js#L1338
As a consequence, the quantity will be added in the first
line and we will have a qty_done of 2 in the first line
and a qty_done of 0 in the second line.
Which will lead to the creation of a back order

opw-5226328

closes odoo/enterprise#101508

X-original-commit: 6cf9f7192e3278042df707d2bd4c53a8a5ced9c4
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Pierre-Louis Hance (plha) <plha@odoo.com>