[FIX] account_reports: Dont refresh report UI with previous options
This fix a traceback:
- Create a report "A" with Period Comparison set to True
- Create a report "B" with Period Comparison set to False
- Create a composite report with report "B" inside.
- Use the variant filter to switch from report "A" to the composite report

This happens because of a UI refresh in `loadReportOptions` trigged by:
- delete this.reportOptionsMap[cacheKey];
- delete this.loadingCallNumberByCacheKey[cacheKey];

When refreshing the UI, it will start with the cachedOption set to the value of the displayed report (report "A"), and is able to pass the condition inside `filters.xml`: `<t t-if="controller.filters.show_period_comparison">`.

While this is happening, the loading of the new report "composite" reaches `this.cachedFilterOptions = options;`. However, the UI refresh is still ongoing, and when it reaches `<t t-if="controller.cachedFilterOptions.comparison.string">` in `filter_comparaison.xml`.
If the new report doesn't have the comparaison filter enabled, cachedFilterOptions.comparison will be undefined triggering the traceback.

This is fixed by marking the objects, that would trigger this useless refresh, as raw preventing them from triggering a refresh.

This only happened on composite reports since the cacheKey used to load them is different from the one received by the call to get_options due to the sections.

task-5085985

closes odoo/enterprise#99780

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>