[FIX] account_batch_payment: don't show warning banner if linked employee has an address
Doing batch payment for sepa payment would generate a warning banner if the partner has no address ( city and country )
However in reality ( already working ) the xml report will be generated with the linked employee address in the case of absence
of the partner address thus it should not show a warning.
The change removes the warning in this case.

closes odoo/enterprise#101042

Task: 5266346
X-original-commit: a6ed1bdfb1da81389fcd7b457219cc8647dea6b7
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>