[FIX] ai: show 'copy to clipboard' on message action of non-ai convo
Before this commit, "Copy to Clipboard" was sometimes not visible in
the message action list when message had some text content.

This happens because message actions can condition overridden by
some other modules, and the way overrides of action condition work is
to have boolean value taking precedence over the local condition of
action.

This architecture is meant to provide exhaustive list of actions to
allow, which is quite useful to control what livechat visitors can
have.

However this architecture requires overrides to properly return
`undefined` when the local condition should apply, which was not the
case for an override by `ai` module. So when `ai` module is
installed, the "copy to clipboard" feature as not available.

This commit fixes the issue by adapting the condition in `ai` module
to return `undefined` on non-AI conversations, and add test coverage
for both `@mail` discuss suite and all overrides of discuss in
`@test_discuss_full_enterprise`.

opw-5240768

closes odoo/enterprise#99794

Related: odoo/odoo#236295
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>