[FIX] hr_payroll: avoid unwanted tracked changes
We use a hack to force the tracking of some fields on the employee.
It causes some problem when you work with the salary configurator as we don't want the changes logged in this case.
The solution is to not track prepare the fields if there is a tracking disable in the context, like mail does.

closes odoo/enterprise#99800

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>