[FIX] l10n_be_reports: prevent error on record not found
This error occurs when the user deletes the partner and then tries to install
the new language.

Steps to reproduce:
- Install 'l10n_be_reports' and switch to 'BE company COA'.
- Open 'res.partner' and delete 'Collection Center - Advance Payment Service'.
- Try to install another language.

Traceback:
ValueError: External ID not found in the system: l10n_be_reports.partner_centre_de_perception_belgium

This error appears after this [commit] because it references
'partner_centre_de_perception_belgium'. If that record is deleted,
an error is raised..

[commit]: https://github.com/odoo/enterprise/commit/b5a935add7090d7da8daee28618f08981de24613

sentry-7039699174

closes odoo/enterprise#99872

Signed-off-by: Ali Alfie (alal) <alal@odoo.com>