[REV] account:  hide sections and notes in journal items tab
This reverts commit 17d0e67106a30a46d608331680e5094dbc44e2e0.

The commit is reverted because the `journal_line_ids` field
is causing issues with onchange methods that rely on cached values.
Specifically, the automatic computation of `debit`/`credit`
when adding new lines to a journal entry was failing.

While `journal_line_ids` (as a subset of `line_ids`) works
correctly when the data is stored in the database,
its absence during an onchange computation
(which relies solely on cache) led to incorrect behavior.

no-task

closes odoo/enterprise#100090

X-original-commit: b34d3b34b35c99734e98a0e644c2e61f40f58c6b
Related: odoo/odoo#236785
Signed-off-by: William André (wan) <wan@odoo.com>