[FIX] l10n_hr: use atn rules in salary configurator
The payslip date is in the current month of when the salary configurator
is used which may be different from the date on the version. That would
lead to have the payslip outside of the contracts date and not take the
ATN into account.

To avoid the issue, we are checking that the we are not in the salary
configurator before returning the value.

We've also add the correct condition on the second ATN records that
didn't check if we are outside the contract

closes odoo/enterprise#99909

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>