[FIX] pos_online_payment*: send to PDIS even if no pay confirmation page
pos_online_payment* = pos_online_payment_self_order_preparation_display
Task: [#5217268](https://www.odoo.com/odoo/project/1737/tasks/5217268)

---

Previously, when an online payment was made, the user was supposed to be redirected to a payment confirmation page which, once the transaction succeeded, sent the related order to the kitchen (PDIS).
However, in some cases, the user never reaches this page.
For example, the user may see the payment succeed in their banking app and close the tab before the redirection happens.

For POS self-orders, we must send the order to the kitchen as soon as the payment is confirmed to avoid confusion between the customer, the cashier, and the kitchen staff.
Relying solely on the confirmation page was therefore unreliable.

---

To fix this, we now leverage the cron that post-processes payment transactions:
we gather all transactions made in self-order or kiosk mode that are not yet post-processed, and send their corresponding orders to the kitchen.
This ensures that orders reach the PDIS even when the user never lands on the confirmation page.

closes odoo/enterprise#101575

X-original-commit: c5fdfaddc64aff8a34ae29f6c00fe5daf7529f23
Related: odoo/odoo#239051
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Ludo Traina (ltra) <ltra@odoo.com>