[FIX] account_accountant: edit partner
Before this commit, when having multiple lines in a statement line.
By default, all the lines have a partner since we compute the partner depending
of other lines. But in the case of an edit or a delete we don't want to
recompute.
What was happening is that the partner keep getting recomputed on the other
lines even the one that was previously edited.

This commit will add a context key when using the reconcile button to recompute
the partners.

opw-5179670

closes odoo/enterprise#100253

X-original-commit: 48e82ee3fcda22da95644c356ecab517c12b44d7
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>