[FIX] hr_payroll: Fix Accounting Section in Payroll Settings
Before this commit, the accounting section in V19 was exclusive to SEPA countries, causing some fields located in this section for other localizations to disappear.
This was the case for the HK localization, which had the Payroll HSBC Autopay option inside this section, causing it not to appear, and also prevented the Batch account move lines option from being available for non-SEPA countries.
With this commit, the accounting section and the Batch account move lines option are available for all localizations.
This was already fixed in the master here is the PR in question enterprise#96283

Also made sure that HSBC autoplay field is only visible in the HK localisation.

task - 5268780

closes odoo/enterprise#100004

Signed-off-by: Abdelrahman Mahmoud (amah) <amah@odoo.com>