[FIX] account_reports: group account basic
Some report must be accessible by user with the invoicing and banks security
group. The menu items are there but when clicking on it, an access error is
thrown.
This commit will add the security rules for the group account basic for
horizontal group (for the partner ledger and aged receivable/payable) and the
tax unit (for the tax report)

opw-5218948

closes odoo/enterprise#100020

Signed-off-by: Julien Alardot (jual) <jual@odoo.com>