[FIX] l10n_es_reports: include NIF in boe export for model 347
The NIF must be included in the BOE export for modelo 347
https://sede.agenciatributaria.gob.es/static_files/Sede/Disenyo_registro
/DR_300_399/archivos/347.pdf pages 3 & 12.
Before this commit, we read the vat but if it doesn't start with 'ES' we
return an empty string because we based on TIN. Now we'll read the vat
(without 'ES' if it starts with it). The user is responsible to fill a
correct number in the vat field of the company.

opw-5207241

closes odoo/enterprise#101290

X-original-commit: 2be8026faafda329abd7df5ee4762ea1efb45004
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>