[FIX] {account,l10n_it}_reports: journal report pdf template
The custom template for the report "Libro Giornale" was neither
inheriting with primary neither using conditions on the country code.
Therefore, the xpaths applied were for all the journal reports,
whatever the company. For example, when exporting the regular journal
report, the names of the journals no longer appeared once the module
l10n_it_reports was installed on a database.

opw-5217520

closes odoo/enterprise#100233

X-original-commit: 52c7210ea5464a78025264abfc579b729f051ca3
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>