[IMP] knowledge: hide GIF and canned response actions
The GIF picker was previously hidden in Knowledge, but became visible again
after the composer actions refactor.

Commit that introduced the change:
https://github.com/odoo/odoo/commit/5e1daafeac231c333ea13369fc208627213d67d4

This commit restores the intended behavior by hiding the GIF action again
and also hides the canned response action, as neither feature is relevant
when composing content in Knowledge, by applying conditions on the
corresponding composer actions.

Task-5163888

closes odoo/enterprise#102927

X-original-commit: 3e5f732d506302af955982f1e5182fa7bbc987d9
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Akshat Trivedi (aktr) <aktr@odoo.com>