[FIX] account_transfer: automatic transfer account rule
Steps to reproduce:
1. Create a new transfer record
2. Add more than one account to the Accounts field
or add a rule with condition on Account
or remove account from an Account condition with multiple accounts set
... (various other actions when adding/removing accounts)

Before:
Redundant and incorrect rules are created and an incorrect value is assigned for account_ids due to the faulty sync between the two.

After:
Removing the onchange mechanism that creates the rule with condition on Account whenever an account is added to the Accounts field, this information is redundant for the user. Also removing the onchange mechanism of the reverse (adding accounts to the Accounts field when a rule with condition on Account is added) as it cannot account for nested rules and any/all conditions. 

opw - 5160635

closes odoo/enterprise#100054

Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>