[FIX] account_iso20022: restore NURG SvcLvl for non-SEPA payments
Since commit [[1]], the Service Level is set to NURG only when using the
specific `iso20022_se` payment method.

However, the previous expected behavior was to set the Service Level to
NURG automatically whenever a payment was in a non-EUR currency or
targeted a non-EU IBAN, regardless of the specific ISO20022 variant.

This commit restores the logic to set the Service Level to NURG for all
standard ISO20022 payments. This fix doesn't check anymore if a payment
is in non-EUR currency or target a non-EU IBAN. Using iso_20022 activate
it by default.

opw-5095483

[1]: https://github.com/odoo/enterprise/commit/67593e5ff9b3a5187a1535bc8fc89590b4c9401e

closes odoo/enterprise#100270

X-original-commit: 72c8edfc6f444f9327c2e239e525ce7741f2f7b2
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>