[FIX] l10n_fr_reports: unbalanced Balance Sheet when coming from the 2024 CoA
odoo/odoo@8f3a869 introduced a new version of the French CoA, legally mandatory starting in 2025. Doing so, it also adapted the P&L and BS reports accordingly. However, it did not take into account the fact that some deprecated account codes would disappear from the P&L, causing the BS to be unbalanced when computing the retained earnings (by calling the P&L with a forced date_scope to run it on the full history).

We fix that by reinjecting the balance of the missing Income and Expense accounts in the computation of the BS's Retained Earnings line.

opw-5212801

closes odoo/enterprise#100239

X-original-commit: 55cea3f8531c186b66a2fd6a1ad5ba49839fe663
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>