[FIX] social_facebook: fix the computation of the stats
Bug
===
Facebook deprecated some of the endpoints related to statistics
https://developers.facebook.com/docs/platforminsights/page/deprecated-metrics

For the accounts, `page_fan_adds` and `page_fan_removes`, we are still unsure
about unfollow, and so to get a fix quickly, we set the audience trend to zero,
and we will fix it later.

For live post, we used `post_impressions` in the API call, but that
metric was not used, so we just remove it.

For stream post, `post_impressions` has been replaced by `post_media_view`,
as suggested in the migration note from Facebook.

Task-5346965

closes odoo/enterprise#100258

X-original-commit: 5e1d0e715c77902f37a40418946be29c3a183766
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>