[FIX] pos_settle_due: reconcile only with partners
Invoice lines to be reconciled are not tailored to the single partner
while the session lines do.
That causes reconciliation to happen between different partners.

To reproduce:
Create an invoice of 100 to Customer 1 and an invoice of 200 to Customer 2.
In a pos session, go to Customer 1, settle 50 on that invoice.
Go to Customer 2, settle 100 on their invoice.
Close register.
=> Lines from different customer are reconciled together.

task-5346644

closes odoo/enterprise#100521

X-original-commit: f09ddde706c2f3c496097e3079588b16ab2b8845
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>