[FIX] documents: issue when creating a new folder and performing any operation
How to reproduce:
=====================
- Select a document
- Click on 'Actions'
- Click on 'Duplicate' (or any other action)
- Create a new folder
- Click on 'Duplicate in My Drive' (or choose any location)

Before this commit:
=====================
The operation was not executed, and it resulted in the creation
of incorrect files.

Technical reason:
=====================
An undefined ID is passed during folder creation, which causes the issue.

After this commit:
=====================
Operations performed while creating a new folder now work correctly
and no incorrect files are generated.

Task-5149773

Part-of: odoo/enterprise#98777
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>