[FIX] voip: mockUserAgent usage
The `mockUserAgent()` is meant to be used with a "platform" ("mac",
"windows", "android"...) as parameter and not a whole user agent string.

In specific cases, a custom string can be used instead, but only to be
added to the user agent string.

This commit adapts its usage(s) accordingly.

closes odoo/enterprise#100154

Related: odoo/odoo#236874
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>