[FIX] l10n_{be,nz,uk}: Remove compute_fiscalyear_dates from _get_start_date_elements
compute_fiscalyear_dates using date.today() is wrong and can lead to different start_dates depending the current date.

For instance if we create a custom fiscal_year from sept 2025 to Dec 2025 and we are currently the 10th Sept 2025.
Then, we try to generate a return for January.
The start date would be set to the 1st Sept.

Then later, when trying to generate the same return for January 2025 but being after that custom fiscal year (for instance 20th January 2026), We get a completely different start date.

The fix is to avoid using the custom fiscal years in get_start_date_elements. An improvement will be done in master to add that feature.

We can also safely remove the custom start_date from the belgian localization as now it is not used anymore.

In 19.0 all of this is removed from this commit: https://github.com/odoo/enterprise/commit/64d690d965a105b9088b382254bab997c85bb594

closes odoo/enterprise#101261

See: https://github.com/odoo/enterprise/pull/100034
X-original-commit: bd920fa720347184f039346d5f07f57b554655dd
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>