[FIX] l10n_nz_eft: Remove commas from payment memos for EFT files
The EFT file format BNZ is exported as a text file. As such, any
commas in the field contents will be treated as a new field,
resulting in incorrect formatting and a rejected batch file.

This behavior is due to the Odoo community commit 3082d3b, which
partially reversed an improvement to the group payment memo field.

- The original behavior was to concatenate the included invoices with
spaces (e.g. 'INV/123 INV/124').
- The improvement replaced this with
the batch payment reference (e.g. 'BATCH/IN/002').
- The behavior after 3082d3b is to concatenate the included
invoices with a space and a comma (e.g. 'INV/123, INV/124'

This commit strips the commas from all generated EFT files, instead
of further altering the group payment memo field. It also alters
existing unit tests to check that the commas are removed.

opw-5099220

closes odoo/enterprise#101215

X-original-commit: 60b66b55bac422cf8074f80af17ed06e01b8e7bb
Signed-off-by: Jacob Was (jacw) <jacw@odoo.com>
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>