[FIX] l10_be_hr_payroll: move hide_in_offer usage to l10n_be_hr_contract_salary
A runbot build error was reported for one of my tests:
https://runbot.odoo.com/odoo/runbot.build.error/233591

The issue was because the field `hide_in_offer` was used in a module
where it did not exist.

I have overridden the function to make it call the field when it was
installed, and moved the logic using it to that inheriting function.

task-5350522

closes odoo/enterprise#100986

X-original-commit: 51e95d4d4836a1ba4d06db9f6b5d4ab02e1ba03a
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>