[FIX] l10n_hk_hr_payroll: Contract start date
There are plenty of places in the code where we check for
the start of a contract; but in all cases we care about
the start of employment.
This corrects the calculations by using the _get_first_version_date()
method to get the actual start of employment.

Part-of: odoo/enterprise#98144
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>