[ADD] l10n_hk_hr_payroll_empf
MPF is a central part of the Hong Kong labor law.
In 2025, they introduced a new central platform (eMPF)
on which all MPF funds are being rolled out.

The aim of this platform is to provide a standardized
way for companies to report their employees MPF contributions
without the need to implement separate file and/or api
integration.

This first step provides the framework to support this platform,
as well as the possibility to export a csv file that can be manually
uploaded on the portal to report the MPF contributions.

API integration will be added in a separate task.

closes odoo/enterprise#98144

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>