[FIX] hr_expense_stripe: Prevent going back on shipping status
It was possible that a card shipping_status would have a wrong value when 2 card.updated event would switch orders.

To replicate:
- Create a physical card
- Order it
- Stop the database from running
- Go on the Stripe Dashboard and simulate the shipping status of the card to "shipped"
- Start back the database
- Go on the Stripe Dashboard and simulate the shipping status of the card to another state such as "delivered"
- The card shipping status will be "delivered"
- Go on the Webhook tab on the Stripe Dashboard and send the failed event from the first notification
- The card shipping status will be "shipped"

Since it's impossible for a shipping status to go back to a previous state, we can skip the update if a previous state is given by stripe.

We also stop sending the shipping address when we are trying to activate the card to prevent any wrong shipping value from blocking this step.

Part-of: odoo/enterprise#95523
Related: odoo/odoo#233590
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>