[FIX] project_enterprise: manage user access for project template
---
## Short functional explanation of the error
Let's say we create a project template with field service. This will enable the start and end date of the project template and its tasks. If we assign an employee, their schedule will be retrieved to see if it matches the start and end date of the project.
If we try to create a project based on this template as an employee who's not an admin (only has rights for the project and services apps, for example), this will result in the error

This operation is allowed for the following groups:
        - Employees/Administrator
        - Employees/Officer: Manage all employees

Contact your administrator to request access if necessary.

Explanation: as we try to retrieve the employee schedule with an employee who doesn't have access rights to the Employee infos, it raises an access error.

## Reproduction Steps
1. As an admin, go to Projects and create a new project with a new stage and a new task. Assign an employee to this task.
2. Enable debug mode and go to the settings of the project. Under field service, check Field Service, then save.
3. Click on the small cog next to the project name and click Convert to Template.
4. Set a planned start and end date for the project and for its task. Assign a customer on the task.
5. Go to Users and select a different user than the one you assigned to the task.
6. Set this user administrator for Project and Field Service.
7. Make sure that this user also has an employee profile. Log in with this user and create a project based on this

### Expected behavior
The project is created without the entries the employee doesn't have access to

### Unexpected behavior
An access error is raised

## Origin of the issue
A sudo() was missing in the code.

__
opw-5263024

closes odoo/enterprise#100055

Signed-off-by: Maeva de Keyser (mdek) <mdek@odoo.com>