[FIX] account_external_tax, l10n_br_avatax: use purchase taxes on bills
Steps to reproduce:
- Install accountant, l10n_br and l10n_br_avatax
- Switch to a Brazilian company
- In Accounting settings, configure AvaTax (require credentials)
- Create a bill
- Compute taxes with AvaTax

Issue:
If a tax like "COFINS Incl." should be added, the Sales tax is added
instead of the Purchases one, resulting in an incorrect account for
the tax line.

Cause:
Brazilian localization allows to use AvaTax for vendor bills, but
the external tax feature doesn't take into account the type of the
tax when searching for one.
It just returns the first one based on the name, the amount and
some other domains.

Solution:
Add the tax type in the domain when searching a Brazilian tax.

opw-5044423

closes odoo/enterprise#100854

X-original-commit: 4bd7a56460fae03ce7aa739eecf35ee290d4d2a0
Related: odoo/odoo#237907
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>