[IMP] pos_iot,pos_blackbox_be: use_lna support for IoT requests
Since odoo/odoo#235702, there is a `point_of_sale.use_lna` system
parameter. When it is set, ePOS requests will use HTTP instead of HTTPS,
and the `targetAddressSpace: "local"` option is used in the `fetch`
request. This bypasses the need for a HTTPS certificate.

This commit adds the same functionality to all IoT requests from the
POS. This should allow the IoT box to function correctly without a HTTPS
certificate.

In addition, we now show a warning message when a request to the
Blackbox via HTTP fails and has to fallback to the websocket.

task-5353672

closes odoo/enterprise#100640

X-original-commit: 4771f98390b38ea31d900a512f0bd2d2f969c3d7
Related: odoo/odoo#237598
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>