[FIX] website_appointment_sale: only delete bookings
Problem: When the module is installed and there are bookings
with answers linked to sale orders, the order lines are deleted
before the order, causing the wrong error message to be displayed.

Solution: The order lines should not be manually unlinked from the
sale order. It should be the bookings linked to the order lines
that should be unlinked before the sale order deletion.

Expected Behavior: When deleting a sale order whose lines have bookings,
the following error message shoud be displayed:
"You can not delete a sent quotation or a confirmed sales order. You must first cancel it."

Current Behavior: When deleting a sale order whose lines have bookings,
the following error message is displayed:
"Once a sales order is confirmed, you can't remove one of its lines
(we need to track if something gets invoiced or delivered).
Set the quantity to 0 instead."

Steps to reproduce on Runbot:
1. Install Appointments and ecommerce
2. For the dental care appointment, edit and enable "Up-front payment"
3. On the website, book an appointment for dental care and proceed to payment
4. Delete the sale order with the booking
5. Observe the incorrect error message being displayed

opw-5092349

closes odoo/enterprise#102747

X-original-commit: 6656e8ef9e05b3ce7ad2d0ef7eaacd80f3c65689
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>