[FIX] project_enterprise: Gantt view with flexible calendar
To reproduce:
=============
1. Activate flexible working hours on the company calendar
2. Go to Project app -> all tasks
3. Switch to Gantt view
notice that on the unassigned tasks row, two days are grayed out randomly

Problem:
========
When flexible working hours is activated, the unavailability intervals
are computed with an estimation: we take the total of working hours
per week and we divide it by number of hours per day to get `N` days, so
we estimate that from `now` to `now + N` days are working days, and the rest
are non-working days. This is wrong because the working hours may not
be evenly distributed over the week, and between today and tomorrow we get
different unavailability intervals.

Solution:
=========
When flexible working hours is activated, we consider that all days
are working days, so there is no unavailability interval to consider.

opw-5257081

closes odoo/enterprise#100548

X-original-commit: 847b55f3aca853d369a899c7830f35e6f4591304
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>