[FIX] sale_subscription: fix invoice description when using quotation template
Issue:

In this issue, when a quotation template is used in sale order,
the end line is missing between name and description in the
invoice pdf.

Steps to reproduce:
1- Install a db with sale and invoicing installed
2- Create a quotation template, and add a description in the line.
3- Create a quotation with the created template.
4- Confirm the order and create an invoice for the sale order.
5- Print the pdf, as seen the name and description are shown in the
   same line, while if the sale order was created without a template,
   we would have seen the description from product in the next line.

`_get_journal_items_full_name` is used to enforce the product name
in the description. However, in cases where product name doesn't
exist in the description, e.g. our case here, the name and
description will be shown in the same line in the invoice pdf.
This can be fixed by adding a new line between the name and the
description.

opw-5130171

closes odoo/enterprise#102308

X-original-commit: 4cda4262dbe1413ad0cba6a17fe59d983a247774
Related: odoo/odoo#240315
Signed-off-by: Mohammadmahdi Alijani (malj) <malj@odoo.com>