[FIX] account_accountant: fix bank statement creation without related accounts
Steps to Reproduce:
1. Install the Accounting module without demo data.
2. In "Chart of Accounts", change the type of all accounts (e.g.;  Expenses).
3. In "Bank" Journal, create a new bank statement line and try to save it.

Error:
SyntaxError - syntax error at or near ")"
LINE 19:                            AND aml.account_id IN ()

Cause:
A IN condition is evaluated with an empty tuple `AND aml.account_id IN ()`.
This is due to that there are no account_ids.

Fix:
This commit ensures that the SQL query is executed only when there are
remaining statement lines and valid account IDs to process. If either is missing,
the method now returns early and updates cron_last_check.

sentry-7059353053

closes odoo/enterprise#100931

X-original-commit: 5ec0af4952bfc395c611a362b156f5681614778e
Signed-off-by: William André (wan) <wan@odoo.com>