[FIX] l10n_us_check_printing: move overlapping date in check print
In the Accounting app, users can print checks created to
pay Vendor Bills. When using any of the US Checks Layout (from
`l10n_us_check_printing`), the payment date would overlap
the table beneath it when the journal's manual sequencing option was
disabled.

This occured because the `.summary_line` container collapsed to a height
of 0 whenever its first child wasn't displayed, as the remaining child
elements are absolutely positioned and therefore do not contribute to
the parent's height.

https://github.com/odoo/enterprise/blob/f54585f84b0fa7a73efb3f0e14266972d510f0a6/l10n_us_check_printing/report/print_check.xml#L24-L34

By assigning a minimum height to the `.summary_line` class, we make sure
that it always occupies space above the table, even when only the date
is shown.

Steps to reproduce:
**/!\ To reproduce the bug locally, you must have a version of
wkhtmltopdf with patched qt. Use `wkhtmltopdf -V` and make sure that
_"(with patched qt)"_ is specified after your wkhtmltopdf version.**
1. Install Accounting (`accountant`) and US Checks Layout (`l10n_us_check_printing`)
2. In Settings > Accounting, toggle on *Checks* if it is not, and
set *Check Layout* to any "US" layout.
3. Go to Accounting > Vendors > Bills
4. Create a new Vendor Bill:
    - Set any Vendor
    - Set any Bill Date
    - Add any Product (with a positive price, the total price of
    the bill must be positive)
    - Confirm
5. Click *Pay*, set the *Payment Method* to *Checks* and click *Create Payment*
6. Click on the *Payments* smart button
7. Click *Print Check* > *Print* (> *Continue*)
8. In the generated PDF, the date overlaps the table beneath it.

opw-5165112

closes odoo/enterprise#100817

X-original-commit: e4d8dc61b34e9e7f57e0e7e9f66ace27969898e6
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>