[FIX] account_accountant: Fix st_line edit with exchange diff
When you create a statement line with one currency rate, and you reconcile it with
a move with a different currency rate, this creates an exchange move.
But when you want to edit the statement line amount, like marking the invoice as
fully paid, this raise a UserError, as the Exchange move is reverted and reconciled,
which means it throw an error like "You are trying to reconcile some entries that are already reconciled."

Step to reproduce:
1 - Create a new currency rate lower than the previous one (ex: 1.5 -> 1.4)
2 - Create a move with the old rate
3 - Create a st_line with the new rate, with an amount a bit lower than the move amount
4 - Open the edit line, mark the invoice as fully paid
5 - Click "Save"
6 - This should raise an UserError("You are trying to reconcile some entries that are already reconciled.")

This is because when we mark the invoice as fully paid, the current exchange diff is reverted and reconciled,
and a new one is created. But when we try to re-reconcile the invoice line with the st_line, the exchange diff
line is still there, and it raise the error.

After this commite: before trying to re-do the reconcilitation between the invoice and the st_line, we filter
out the exchange diff, as it shouldn't interfer with the reconcilitation process.

opw-5184679

closes odoo/enterprise#100864

X-original-commit: c63fa6f
Related: odoo/odoo#237931
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>