[FIX] hr_contract_salary: refresh new offer page with correct context
Steps to Reproduce
==================
- Go to Employees
- Choose an employee with no offers
- Click on the "Offers - new" smart button (a form view will open with the correct employee name at the bottom)
- Refresh the offer's page without saving (the employee field is emptied and the required Applicant field appears)

Issue
==================
The generate offer action that is triggered through the smart button returns a one-time action dictionary with the context.
But that context is lost when we refresh the offer page as it's not saved anywhere.

Fix
==================
Replace the action dictionary with an action record for creating a new offer to ensure the context is maintained upon page refresh.

closes odoo/enterprise#102260

Task-id: 5059490
Signed-off-by: Abdelrahman Mahmoud (amah) <amah@odoo.com>