[FIX] sale_subscription: ensure next invoice increase for free products
Before this commit, when a product invoiced based on delviered quantity
was not delivered, the subscription cron would detect the free period
but the next invoice date would not be incremented. As a result, the
cron would run every day and post a 'Automatic renewal succeeded. Free
subscription.' message in the chatter.

task-5345944

closes odoo/enterprise#100571

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>