[FIX] account_accountant: prevent error when setting partner on statement line
Currently, an error occurs when a user sets the partner on a statement line.

Steps to Reproduce:
 - Install the account_accountant module.
 - Go to Invoices and create two invoices with the same partner and the same
   total amount by adding an invoice line.
 - Go to the Dashboard and click Last Statement under the Bank journal.
 - Create a new statement line with the same amount as the invoice amount.
 - Click Set Partner and select the same partner used in the invoices.

TypeError: '<=' not supported between instances of 'bool' and 'datetime.date'

After this commit, when a user sets the partner on the statement line,
the system attempts to automatically reconcile the line. It then searches
for a single invoice matching the statement line amount [1] and tries to
select the one with the closest prior or equal date. However, draft
invoices have not an invoice_date, and when the system attempts
to filter on this field, and the error is raised [2].

This commit ensures that the comparison between invoice date and
statement date occurs only when the invoice date is present.

[this comit]; https://github.com/odoo/enterprise/pull/98269

[1]; https://github.com/odoo/enterprise/blob/84ccb807852ab6b118466d12c8f02ea4da7fa3db/account_accountant/models/account_bank_statement.py#L585

[2]; https://github.com/odoo/enterprise/blob/84ccb807852ab6b118466d12c8f02ea4da7fa3db/account_accountant/models/account_bank_statement.py#L249

sentry-7063516380

closes odoo/enterprise#100576

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>