[FIX] account_bank_statement_import_csv: init model
When importing a bank statements xlsx in a bank journal, the wizard would
not match the "Cumulative Balance", not even proposing it for manual setup.

Steps to reproduce:
- Go to the accounting dashboard
- On a bank journal tile, in the right corner menu "New > Import File"
- Upload a file (there is one attached to the ticket)
- Cumulative Balance is not matched

This commit adds module.init() that is skipped in the "onWillStart"
override (it's present in the parent onWillStart). This has for consequence
that the "bank_stmt_import" key is now present in the context when
get_fields_tree() from Base_ImportImport is called, allowing the addition
of missing field.
See https://github.com/odoo/enterprise/blob/19.0/account_bank_statement_import_csv/wizard/account_bank_statement_import_csv.py#L18

This commit also adds a check to only add debit & credit in the added
field list if they are not actual field on the bst line model (see:
https://github.com/odoo/enterprise/commit/af863c5a53d0ab50fe67cb9ea910391d4a1979dd)
This commit also checks that those fields are only added when the model
is account bank statement line (only usefull in this case).

opw-5222326

closes odoo/enterprise#100693

Signed-off-by: Ali Alfie (alal) <alal@odoo.com>