[FIX] l10n_be_hr_contract_salary: fix Mobility Budget computation in part time
Mobility budget should be computed based on a full time equivalent:
https://lebudgetmobilite.be/fr/6-quel-est-le-montant-du-budget-mobilite#tempspartielTCOcalcul

During salary simulation, that is based on work time rate, a payslip is
computed and triggers to recompute the version on which the payslip is based.
That makes the mobility budget recompute as well, but this time, based on a
part time wage (e.g.: 2000 instead of 4000 for a mid-time).

This commit adds context keys to ensure that during the salary simulation,
the Mobility budget is still computed based on a full time equivalent.

Task-5360844

Part-of: odoo/enterprise#100597
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>