[IMP] l10n_br_avatax: send delivery address for NF-e
Purpose: For invoice goods(NF-e), the delivery address can differ
from the invoice address. It is necessary for both fiscal and
logistical reasons to provide the delivery address in the tax
calculation and EDI requests. Failure to do so may result in
tax issues for the issuer.

Expected Behavior: Users will be able to set a delivery address
on the sales order and invoice and the delivery address will be
sent in the requests for tax calculation and EDI. The returned
EDI document will display the delivery address information.

Current Behavior: The delivery address is not sent in the requests.

task-5171186

closes odoo/enterprise#101267

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>