[FIX] l10n_lu_reports: add PurchaseInvoices tag
Since version 2.01 of FAIA, it is a requirement to include the `PurchaseInvoices` tag in the FAIA XML.

With a `l10n_lu` company:

- Create a vendor bill.
- In the general ledger, download the FAIA XML report.

In the generated XML document, the vendor bill will not appear under `PurchaseInvoices`.

The `PurchaseInvoices` section is meant to mirror the `SalesInvoices` section and contains the exact same structure.
This PR implements a generic template for both.

**Specs and source** (in `FAIA_v2.01_full`):
https://pfi.public.lu/dam-assets/backup/FAIA/FAIA/XSD_Files.zip

opw-5360572

closes odoo/enterprise#102180

X-original-commit: 55680adfc454e0ee69c27320e1970df9193bc935
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>